requirements.txt
aioboto3==13.4.0
aiobotocore==2.18.0
aiofiles==24.1.0
aiohappyeyeballs==2.6.1
aiohttp==3.11.16
aioitertools==0.12.0
aiosignal==1.3.2
alembic==1.13.1
annotated-types==0.7.0
anthropic==0.49.0
anyio==4.9.0
async-property==0.2.2
atlassian-python-api==4.0.4
attrs==25.3.0
banks==2.1.2
beautifulsoup4==4.13.4
boto3==1.36.1
botocore==1.36.1
cached-property==1.5.2
cachetools==5.5.2
certifi==2025.1.31
cffi==1.17.1
charset-normalizer==3.4.1
click==8.1.8
colorama==0.4.6
cryptography==44.0.2
dataclasses-json==0.6.7
Deprecated==1.2.18
deprecation==2.1.0
dirtyjson==1.0.8
distro==1.9.0
dnspython==2.7.0
docopt==0.6.2
ecdsa==0.19.1
email_validator==2.2.0
fastapi==0.115.12
fastapi-cli==0.0.7
filetype==1.2.0
frozenlist==1.6.0
fsspec==2025.3.2
google-auth==2.39.0
greenlet==3.2.0
griffe==1.7.2
h11==0.14.0
html2text==2025.4.15
httpcore==1.0.8
httptools==0.6.4
httpx==0.26.0
idna==3.10
itsdangerous==2.2.0
Jinja2==3.1.6
jiter==0.9.0
jmespath==1.0.1
joblib==1.4.2
jwcrypto==1.5.6
llama-cloud==0.1.18
llama-cloud-services==0.6.12
llama-hub==0.0.45
llama-index==0.12.31
llama-index-agent-openai==0.4.6
llama-index-cli==0.4.1
llama-index-core==0.12.31
llama-index-embeddings-openai==0.3.1
llama-index-indices-managed-llama-cloud==0.6.11
llama-index-llms-anthropic==0.6.10
llama-index-llms-bedrock==0.3.8
llama-index-llms-bedrock-converse==0.4.15
llama-index-llms-openai==0.3.37
llama-index-multi-modal-llms-openai==0.4.3
llama-index-program-openai==0.3.1
llama-index-question-gen-openai==0.3.0
llama-index-readers-file==0.4.7
llama-index-readers-llama-parse==0.4.0
llama-index-tools-tavily-research==0.3.0
llama-parse==0.6.12
Mako==1.3.10
markdown-it-py==3.0.0
MarkupSafe==3.0.2
marshmallow==3.26.1
mdurl==0.1.2
multidict==6.4.3
mypy-extensions==1.0.0
nest-asyncio==1.6.0
networkx==3.4.2
nltk==3.9.1
numpy==2.2.5
oauthlib==3.2.2
openai==1.75.0
orjson==3.10.16
packaging==25.0
pandas==2.2.3
pillow==11.2.1
pipreqs==0.4.13
platformdirs==4.3.7
propcache==0.3.1
psutil==7.0.0
psycopg==3.2.9
psycopg-binary==3.2.9
pyasn1==0.6.1
pyasn1_modules==0.4.2
pycparser==2.22
pydantic==2.11.3
pydantic-extra-types==2.10.3
pydantic-settings==2.9.1
pydantic_core==2.33.1
Pygments==2.19.1
pypdf==5.4.0
python-dateutil==2.9.0.post0
python-dotenv==1.1.0
python-keycloak==5.3.1
python-multipart==0.0.20
pytz==2025.2
PyYAML==6.0.2
regex==2024.11.6
requests==2.32.3
requests-oauthlib==2.0.0
requests-toolbelt==1.0.0
retrying==1.3.4
rich==14.0.0
rich-toolkit==0.14.1
rsa==4.9.1
s3transfer==0.11.3
shellingham==1.5.4
six==1.17.0
sniffio==1.3.1
soupsieve==2.7
SQLAlchemy==2.0.40
sse-starlette==2.2.1
starlette==0.46.2
striprtf==0.0.26
tavily-python==0.7.2
tenacity==9.1.2
tiktoken==0.9.0
tqdm==4.67.1
typer==0.15.2
typing-inspect==0.9.0
typing-inspection==0.4.0
typing_extensions==4.13.2
tzdata==2025.2
ujson==5.10.0
urllib3==2.4.0
uvicorn==0.34.2
uvloop==0.21.0
watchfiles==1.0.5
websockets==15.0.1
wrapt==1.17.2
yarg==0.1.10
yarl==1.20.0
